+++ truffle/src/com.oracle.truffle.api.object.test/src/com/oracle/truffle/api/object/test/DynamicTypeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -58,6 +58,11 @@ public class DynamicTypeTest extends ParametrizedDynamicObjectTest {
     @Parameters(name = "{0}")
     public static List<TestRun> data() {
         return Arrays.asList(TestRun.values());
+    }
+
+    @Test
+    public void testDefaultDynamicTypeIsShared() {
+        assertSame(Shape.newBuilder().build().getDynamicType(), Shape.newBuilder().build().getDynamicType());
     }
 
     @Test